pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.5.0"
description = ""
authors = ["Konstantin Vlasov <ke.vlasov@yandex.ru>"]
readme = "README.md"

packages = [
  { include = "gendiff" },
]

repository = "https://github.com/kvlas/python-project-50"
homepage = "https://github.com/kvlas/python-project-50"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
generate_diff = "gendiff.scripts.generate_diff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
